# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Typing Speed Test API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TypingResult {
  id: number;
  userId: string;
  wpm: number;
  cpm: number;
  accuracy: number;
  duration: number;
  difficulty: string;
  mistakeCount: number;
  /** @nullable */
  passageId?: number | null;
  createdAt: string;
}

export interface TypingResultInput {
  wpm: number;
  cpm: number;
  accuracy: number;
  duration: number;
  difficulty: string;
  mistakeCount: number;
  /** @nullable */
  passageId?: number | null;
}

export interface UserStats {
  totalTests: number;
  bestWpm: number;
  averageWpm: number;
  averageAccuracy: number;
  totalTimeSeconds: number;
  currentStreak: number;
  /** @nullable */
  rank?: number | null;
}

export interface WpmHistoryPoint {
  date: string;
  wpm: number;
  accuracy: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  wpm: number;
  accuracy: number;
  duration: number;
  difficulty: string;
  createdAt: string;
}

export interface Achievement {
  id: number;
  key: string;
  title: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
}

export interface Passage {
  id: number;
  text: string;
  difficulty: string;
  wordCount: number;
}

export interface DailyChallenge {
  date: string;
  passage: Passage;
  leaderboard: LeaderboardEntry[];
}

export type ListResultsParams = {
limit?: number;
offset?: number;
};

export type GetStatsHistoryParams = {
limit?: number;
};

export type GetLeaderboardParams = {
/**
 * Filter by test duration in seconds
 */
duration?: GetLeaderboardDuration;
difficulty?: GetLeaderboardDifficulty;
limit?: number;
};

export type GetLeaderboardDuration = typeof GetLeaderboardDuration[keyof typeof GetLeaderboardDuration];


export const GetLeaderboardDuration = {
  NUMBER_15: 15,
  NUMBER_30: 30,
  NUMBER_60: 60,
  NUMBER_120: 120,
} as const;

export type GetLeaderboardDifficulty = typeof GetLeaderboardDifficulty[keyof typeof GetLeaderboardDifficulty];


export const GetLeaderboardDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ListPassagesParams = {
difficulty?: ListPassagesDifficulty;
count?: number;
};

export type ListPassagesDifficulty = typeof ListPassagesDifficulty[keyof typeof ListPassagesDifficulty];


export const ListPassagesDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;