# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Typing Speed Test API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: number;
  key: string;
  title: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
}