# lib/api-zod/src/generated/types/dailyChallenge.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Typing Speed Test API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';
import type { Passage } from './passage';

export interface DailyChallenge {
  date: string;
  passage: Passage;
  leaderboard: LeaderboardEntry[];
}