# lib/api-zod/src/generated/types/passage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Typing Speed Test API
 * OpenAPI spec version: 0.1.0
 */

export interface Passage {
  id: number;
  text: string;
  difficulty: string;
  wordCount: number;
}