# lib/api-zod/src/generated/types/typingResultInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Typing Speed Test API
 * OpenAPI spec version: 0.1.0
 */

export interface TypingResultInput {
  wpm: number;
  cpm: number;
  accuracy: number;
  duration: number;
  difficulty: string;
  mistakeCount: number;
  /** @nullable */
  passageId?: number | null;
}